Разбивает код на модули
