+++ index.html
@@ -7,8 +7,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
-  <script src="js/functions.js" defer="defer"></script>
-  <script src="js/main.js" defer="defer"></script>
+  <script src="js/main.js" type="module" defer="defer"></script>
 </head>
 
 <body>
