+++ js/util.js
@@ -34,3 +34,5 @@ const getRandomInteger = (a, b) => {
   const result = Math.random() * (upper - lower + 1) + lower;
   return Math.floor(result);
 };
+
+export {getRandomInteger};
